requirements.txt
Pillow==10.2.0
mtcnn==0.1.1
piexif==1.1.3
tqdm==4.66.2
setuptools==69.2.0
tensorflow==2.19.0 